+++ code_examples/sdk_examples/src/core_features/public_credentials/01_create_credential.ts
@@ -1,10 +1,28 @@
 import * as Kilt from '@kiltprotocol/sdk-js'
 
 export function createNftCollectionCredential(
-  ctype: Kilt.ICType,
   assetDid: Kilt.AssetDidUri,
   artistDid: Kilt.DidUri
 ): Kilt.IPublicCredentialInput {
+  // CType definition.
+  const ctype = Kilt.CType.fromProperties(
+    `NFT Collection Certification CType`,
+    {
+      name: {
+        type: 'string'
+      },
+      pieces: {
+        type: 'integer'
+      },
+      creationDate: {
+        type: 'string'
+      },
+      artistIdentity: {
+        type: 'string'
+      }
+    }
+  )
+
   const claimProperties: Kilt.IClaimContents = {
     name: 'Awesome NFT drop',
     // NFT collection only has 100 pieces in total
