+++ code_examples/sdk_examples/src/core_features/public_credentials/02_issue_credential.ts
