+++ code_examples/sdk_examples/src/core_features/public_credentials/03_retrieve_credential_by_id.ts
