+++ code_examples/sdk_examples/src/core_features/public_credentials/04_retrieve_credentials_by_subject.ts
