+++ code_examples/sdk_examples/src/core_features/public_credentials/05_verify_credential.ts
