+++ code_examples/sdk_examples/src/core_features/public_credentials/06_revoke_remove_credential_by_id.ts
