+++ code_examples/sdk_examples/src/core_features/public_credentials/07_revoke_remove_credential_by_content.ts
