+++ code_examples/sdk_examples/src/core_features/public_credentials/08_unrevoke_credential.ts
