+++ code_examples/sdk_examples/src/core_features/public_credentials/09_reclaim_deposit.ts
