+++ code_examples/sdk_examples/src/core_features/public_credentials/index.ts
@@ -2,16 +2,15 @@ import * as Kilt from '@kiltprotocol/sdk-js'
 
 import { createCompleteFullDid } from '../did/05_full_did_complete'
 
-import { createNftCollectionCType } from './01_create_ctype'
-import { createNftCollectionCredential } from './02_create_credential'
-import { fetchCredentialById } from './04_retrieve_credential_by_id'
-import { issueCredential } from './03_issue_credential'
-import { reclaimDeposit } from './10_reclaim_deposit'
-import { retrieveAllAssetCredentials } from './05_retrieve_credentials_by_subject'
-import { revokeCredential } from './08_revoke_remove_credential_by_content'
-import { revokeCredentialById } from './07_revoke_remove_credential_by_id'
-import { unrevokeCredential } from './09_unrevoke_credential'
-import { verifyCredential } from './06_verify_credential'
+import { createNftCollectionCredential } from './01_create_credential'
+import { fetchCredentialById } from './03_retrieve_credential_by_id'
+import { issueCredential } from './02_issue_credential'
+import { reclaimDeposit } from './09_reclaim_deposit'
+import { retrieveAllAssetCredentials } from './04_retrieve_credentials_by_subject'
+import { revokeCredential } from './07_revoke_remove_credential_by_content'
+import { revokeCredentialById } from './06_revoke_remove_credential_by_id'
+import { unrevokeCredential } from './08_unrevoke_credential'
+import { verifyCredential } from './05_verify_credential'
 
 import { generateKeypairs } from '../utils/generateKeypairs'
 
@@ -29,27 +28,13 @@ export async function runAll(
     })
   )
 
-  console.log('1 public credentials) Create CType')
-  const ctype = await createNftCollectionCType(
-    attesterDid.uri,
-    submitterAccount,
-    async ({ data }) => ({
-      signature: keypairs.attestation.sign(data),
-      keyType: keypairs.attestation.type
-    })
-  )
-
-  console.log('2 public credentials) Create credential object')
+  console.log('1 public credentials) Create credential object')
   const { authentication } = generateKeypairs()
   const artistDid = Kilt.Did.getFullDidUriFromKey(authentication)
   const collectionDid: Kilt.AssetDidUri =
     'did:asset:eip155:1.erc721:0xb47e3cd837ddf8e4c57f05d70ab865de6e193bbb'
-  const credential = createNftCollectionCredential(
-    ctype,
-    collectionDid,
-    artistDid
-  )
-  console.log('3 public credentials) Issue credential')
+  const credential = createNftCollectionCredential(collectionDid, artistDid)
+  console.log('2 public credentials) Issue credential')
   await issueCredential(
     attesterDid.uri,
     submitterAccount,
@@ -59,7 +44,7 @@ export async function runAll(
     }),
     credential
   )
-  console.log('4 public credentials) Fetch credential by ID')
+  console.log('3 public credentials) Fetch credential by ID')
   const credentialId = Kilt.PublicCredential.getIdForCredential(
     credential,
     attesterDid.uri
@@ -70,17 +55,17 @@ export async function runAll(
       `Was not possible to retrieve just-issued credential with ID ${credentialId}`
     )
   }
-  console.log('5 public credentials) Retrieve credentials by subject')
+  console.log('4 public credentials) Retrieve credentials by subject')
   const retrievedCredentials = await retrieveAllAssetCredentials(collectionDid)
   if (!retrievedCredentials) {
     throw new Error(
       `Was not possible to retrieve just-issued credentials for asset ${collectionDid}`
     )
   }
-  console.log('6 public credentials) Verify credential')
+  console.log('5 public credentials) Verify credential')
   await verifyCredential(fetchedCredential)
   await verifyCredential(retrievedCredentials[0])
-  console.log('7 public credentials) Revoke and remove credential by ID')
+  console.log('6 public credentials) Revoke and remove credential by ID')
   await revokeCredentialById(
     attesterDid.uri,
     submitterAccount,
@@ -91,7 +76,7 @@ export async function runAll(
     credentialId,
     true
   )
-  console.log('8.1 public credentials) Re-issue credential')
+  console.log('7.1 public credentials) Re-issue credential')
   await issueCredential(
     attesterDid.uri,
     submitterAccount,
@@ -101,7 +86,7 @@ export async function runAll(
     }),
     credential
   )
-  console.log('8.2 public credentials) Revoke credential')
+  console.log('7.2 public credentials) Revoke credential')
   await revokeCredential(
     attesterDid.uri,
     submitterAccount,
@@ -111,7 +96,7 @@ export async function runAll(
     }),
     credential
   )
-  console.log('9 public credentials) Unrevoke credential')
+  console.log('8 public credentials) Unrevoke credential')
   await unrevokeCredential(
     attesterDid.uri,
     submitterAccount,
@@ -121,7 +106,7 @@ export async function runAll(
     }),
     credential
   )
-  console.log('10 public credentials) Reclaim deposit')
+  console.log('9 public credentials) Reclaim deposit')
   await reclaimDeposit(submitterAccount, fetchedCredential)
 
   console.log('Public credentials flow completed!')
